// extension/src/components/errorlens/declares.d.ts
export declare type TELExtensionConfig = Readonly<IELExtensionConfigType>;
// ...
export declare type TELGroupedByLineDiagnostics = Record<
  string,
  vscode.Diagnostic[]
>;
// ...
export declare type TELDecorationKeys =
  | 'error'
  | 'warning'
  | 'info'
  | 'hint'
  | 'gutterError'
  | 'gutterWarning'
  | 'gutterInfo'
  | 'gutterHint'
  | 'errorRange'
  | 'warningRange'
  | 'infoRange'
  | 'hintRange'
  | 'multilineError'
  | 'multilineWarning'
  | 'multilineInfo'
  | 'multilineHint'
  | 'multilineErrorLineBackground'
  | 'multilineHintLineBackground'
  | 'multilineInfoLineBackground'
  | 'multilineWarningLineBackground'
  | 'transparent1x1Icon';
// ...
export declare interface IELGutter {
	iconSet: TELExtensionConfig['gutterIconSet'];

	errorIconPath: vscode.Uri | string;
	errorIconPathLight: vscode.Uri | string;

	warningIconPath: vscode.Uri | string;
	warningIconPathLight: vscode.Uri | string;

	infoIconPath: vscode.Uri | string;
	infoIconPathLight: vscode.Uri | string;

	hintIconPath: vscode.Uri | string | undefined;
	hintIconPathLight: vscode.Uri | string | undefined;

	transparent1x1Icon: vscode.Uri | string;
}